fixed various compile warnings related to 64 bits
git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@3335 689a6050-402a-0410-94f2-e92a70836424

+++ modules_k/perlvdb/perlvdbfunc.c
@@ -136,7 +136,7 @@ db_con_t* perlvdb_db_init(const char* url) {
 		return NULL;
 	}
 	memset(res, 0, consize);
-	CON_TAIL(res) = (unsigned int)obj;
+	CON_TAIL(res) = (unsigned int)(unsigned long)obj;
 
 	return res;
 }
