+++ modules_k/textops/textops.c
@@ -803,7 +803,7 @@ static int filter_body_f(struct sip_msg* msg, char* _content_type,
 		    if (del_lump(msg, body.s - msg->buf, start - body.s, 0)
 			== 0) {
 			LM_ERR("Deleting lump <%.*s> failed\n",
-			       start - body.s, body.s);
+			       (int)(start - body.s), body.s);
 			return -1;
 		    }
 		    if (find_line_start("--Boundary", 10, &start, &len)) {
